version 2.3: - create settings file - Chuyển file profile vào file settings
