+++ app/src/main/AndroidManifest.xml
@@ -67,14 +67,7 @@
             android:name=".controller.user.ChatActivity"
             android:exported="true" />
         <activity
-            android:name=".controller.user.ProfileActivity"
-            android:exported="true" />
-        <activity
             android:name=".controller.SettingsActivity"
             android:exported="true" />
-        <activity
-            android:name=".controller.GeneralSettingsActivity"
-            android:exported="true" />
     </application>
-
 </manifest>
