+++ app/src/main/java/com/lan/campsiteproject/controller/SettingsActivity.java
@@ -1,36 +1,38 @@
 package com.lan.campsiteproject.controller;
 
 import android.os.Bundle;
-import android.widget.AdapterView;
-import android.widget.ArrayAdapter;
-import android.widget.ListView;
-
+import android.widget.ImageButton;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.fragment.app.Fragment;
-
+import com.lan.campsiteproject.controller.ProfileSettingsFragment;
+import com.lan.campsiteproject.controller.GeneralSettingsFragment;
 import com.lan.campsiteproject.R;
 
 public class SettingsActivity extends AppCompatActivity {
-
-    private String[] menuItems = {"Profile", "General"};
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_settings);
 
-        ListView menu = findViewById(R.id.settings_menu);
-        menu.setAdapter(new ArrayAdapter<>(this, android.R.layout.simple_list_item_activated_1, menuItems));
-        menu.setChoiceMode(ListView.CHOICE_MODE_SINGLE);
+        ImageButton btnProfile = findViewById(R.id.btnProfile);
+        ImageButton btnGeneral = findViewById(R.id.btnGeneral);
+        ImageButton btnReturn = findViewById(R.id.btnReturn);
 
-        menu.setOnItemClickListener((AdapterView<?> parent, android.view.View view, int position, long id) -> {
-            Fragment fragment = position == 0 ? new ProfileSettingsFragment() : new GeneralSettingsFragment();
+        btnProfile.setOnClickListener(v -> getSupportFragmentManager().beginTransaction()
+                .replace(R.id.settings_content, new ProfileSettingsFragment())
+                .commit());
+
+        btnGeneral.setOnClickListener(v -> getSupportFragmentManager().beginTransaction()
+                .replace(R.id.settings_content, new GeneralSettingsFragment())
+                .commit());
+
+        btnReturn.setOnClickListener(v -> finish());
+
+        // Load default fragment (Profile)
+        if (savedInstanceState == null) {
             getSupportFragmentManager().beginTransaction()
-                    .replace(R.id.settings_content, fragment)
+                    .replace(R.id.settings_content, new ProfileSettingsFragment())
                     .commit();
-        });
-
-        // Default selection
-        menu.performItemClick(menu.getAdapter().getView(0, null, null), 0, menu.getAdapter().getItemId(0));
+        }
     }
 }
