+++ app/src/main/java/com/lan/campsiteproject/controller/campsite/ListCampsiteActivity.java
@@ -24,6 +24,7 @@ import com.lan.campsiteproject.controller.user.ChatListActivity;
 import com.lan.campsiteproject.model.Campsite;
 import com.lan.campsiteproject.controller.user.LoginActivity;
 import com.google.firebase.auth.FirebaseAuth;
+import com.lan.campsiteproject.controller.SettingsActivity;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -65,9 +66,6 @@ public class ListCampsiteActivity extends AppCompatActivity {
             popupView.findViewById(R.id.btnOrder).setOnClickListener(btn -> {
                 startActivity(new Intent(this, com.lan.campsiteproject.controller.orders.OrderHistoryActivity.class));
                 popupWindow.dismiss();
-            });
-            popupView.findViewById(R.id.btnProfile).setOnClickListener(btn -> {
-                startActivity(new Intent(this, com.lan.campsiteproject.controller.user.ProfileActivity.class));
             });
             popupView.findViewById(R.id.btnLogout).setOnClickListener(btn -> {
                 FirebaseAuth.getInstance().signOut();
