+++ app/src/main/res/drawable/ic_person.xml
@@ -1,0 +1,7 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp" android:height="24dp" android:viewportWidth="24"
+    android:viewportHeight="24">
+    <path
+        android:fillColor="#FFFFFF"
+        android:pathData="M12,12c2.7,0 8,1.34 8,4v2H4v-2c0-2.66 5.3-4 8-4zm0-2a4,4 0 1,0 0,-8 4,4 0 0,0 0,8z"/>
+</vector>
