+++ app/src/main/res/drawable/ic_settings.xml
@@ -1,0 +1,7 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp" android:height="24dp" android:viewportWidth="24"
+    android:viewportHeight="24">
+    <path
+        android:fillColor="#FFFFFF"
+        android:pathData="M19.14,12.94c0.04,-0.31 0.06,-0.63 0.06,-0.94s-0.02,-0.63 -0.06,-0.94l2.03,-1.58c0.18,-0.14 0.23,-0.41 0.12,-0.61l-1.92,-3.32c-0.12,-0.2 -0.37,-0.26 -0.57,-0.19l-2.39,0.96a7.03,7.03 0 0,0 -1.63,-0.94l-0.36,-2.54A0.486,0.486 0 0,0 14,2h-4c-0.24,0 -0.44,0.17 -0.48,0.41l-0.36,2.54c-0.6,0.23 -1.17,0.54 -1.7,0.9l-2.39,-0.96c-0.21,-0.08 -0.45,-0.01 -0.57,0.19l-1.92,3.32c-0.12,0.2 -0.07,0.47 0.12,0.61l2.03,1.58c-0.04,0.31 -0.06,0.63 -0.06,0.94s0.02,0.63 0.06,0.94l-2.03,1.58c-0.18,0.14 -0.23,0.41 -0.12,0.61l1.92,3.32c0.12,0.2 0.37,0.26 0.57,0.19l2.39,-0.96c0.51,0.36 1.07,0.67 1.67,0.9l0.36,2.54c0.04,0.24 0.24,0.41 0.48,0.41h4c0.24,0 0.44,-0.17 0.48,-0.41l0.36,-2.54c0.6,-0.23 1.17,-0.54 1.7,-0.9l2.39,0.96c0.21,0.08 0.45,0.01 0.57,-0.19l1.92,-3.32c0.12,-0.2 0.07,-0.47 -0.12,-0.61l-2.03,-1.58zM12,15.5a3.5,3.5 0 1,1 0,-7 3.5,3.5 0 0,1 0,7z"/>
+</vector>
