+++ app/src/main/res/layout/activity_settings.xml
@@ -1,20 +1,57 @@
-<?xml version="1.0" encoding="utf-8"?>
+<!-- app/src/main/res/layout/activity_settings.xml -->
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="horizontal">
 
-    <ListView
-        android:id="@+id/settings_menu"
-        android:layout_width="200dp"
+    <LinearLayout
+        android:id="@+id/settings_menu_panel"
+        android:layout_width="0dp"
         android:layout_height="match_parent"
-        android:background="#222"
-        android:divider="@android:color/darker_gray"
-        android:dividerHeight="1dp"/>
+        android:layout_weight="2"
+        android:orientation="vertical"
+        android:background="@android:color/darker_gray"
+        android:gravity="center_horizontal">
+
+        <ImageButton
+            android:id="@+id/btnProfile"
+            android:layout_width="48dp"
+            android:layout_height="48dp"
+            android:layout_marginTop="32dp"
+            android:background="?attr/selectableItemBackgroundBorderless"
+            android:src="@drawable/ic_person"
+            android:contentDescription="Profile"
+            android:tint="@android:color/white"/>
+
+        <ImageButton
+            android:id="@+id/btnGeneral"
+            android:layout_width="48dp"
+            android:layout_height="48dp"
+            android:layout_marginTop="24dp"
+            android:background="?attr/selectableItemBackgroundBorderless"
+            android:src="@drawable/ic_settings"
+            android:contentDescription="General"
+            android:tint="@android:color/white"/>
+
+        <View
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            android:layout_weight="1"/>
+
+        <ImageButton
+            android:id="@+id/btnReturn"
+            android:layout_width="48dp"
+            android:layout_height="48dp"
+            android:layout_marginBottom="32dp"
+            android:background="?attr/selectableItemBackgroundBorderless"
+            android:src="@drawable/ic_arrow_back"
+            android:contentDescription="Return"
+            android:tint="@android:color/white"/>
+    </LinearLayout>
 
     <FrameLayout
         android:id="@+id/settings_content"
         android:layout_width="0dp"
         android:layout_height="match_parent"
-        android:layout_weight="1"/>
+        android:layout_weight="8"/>
 </LinearLayout>
